# internal/models/model_auth_token.go
/*
 * Minesweeper
 *
 * Backend API for a minesweeper game app
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

import (
	"time"
)

type AuthToken struct {
	Id string `json:"id,omitempty" bson:"_id,omitempty"`

	Timestamp time.Time `json:"timestamp,omitempty"`
}