# internal/models/model_user.go
/*
 * Minesweeper
 *
 * Backend API for a minesweeper game app
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type User struct {
	Id string `json:"id,omitempty" bson:"_id,omitempty"`

	Username string `json:"username" bson:"username"`

	Email string `json:"email" bson:"email"`

	Password string `json:"password" bson:"password"`
}